Revert "refactor(web): split root file (#3522)"
This reverts commit 2425e6b926149228b017243af843ba302e6932e5.
