+++ mergify_engine/web/root.py
@@ -19,40 +19,45 @@ import typing
 import daiquiri
 from datadog import statsd
 import fastapi
+import httpx
 from starlette import requests
 from starlette import responses
+import voluptuous
 import yaaredis
 
 # To load the json serializer need to read queues
 from mergify_engine import check_api  # noqa
+from mergify_engine import config
+from mergify_engine import count_seats
+from mergify_engine import github_events
 from mergify_engine import github_types
 from mergify_engine import json
 from mergify_engine import utils
-from mergify_engine.clients import github as github_client
+from mergify_engine.clients import github
 from mergify_engine.clients import http
+from mergify_engine.dashboard import subscription
+from mergify_engine.dashboard import user_tokens
 from mergify_engine.queue import merge_train
 from mergify_engine.web import auth
 from mergify_engine.web import badges
 from mergify_engine.web import config_validator
-from mergify_engine.web import dashboard
-from mergify_engine.web import github
 from mergify_engine.web import redis
-from mergify_engine.web import refresher
 from mergify_engine.web import simulator
-from mergify_engine.web.api import root as api_root
+from mergify_engine.web.api import root
 
 
 LOG = daiquiri.getLogger(__name__)
 
 app = fastapi.FastAPI()
-app.include_router(dashboard.router)
-app.include_router(github.router)
-app.include_router(refresher.router)
-
 app.mount("/simulator", simulator.app)
 app.mount("/validate", config_validator.app)
 app.mount("/badges", badges.app)
-app.mount("/v1", api_root.app)
+app.mount("/v1", root.app)
+
+# Set the maximum timeout to 5 seconds: GitHub is not going to wait for
+# more than 10 seconds for us to accept an event, so if we're unable to
+# forward an event in 5 seconds, just drop it.
+EVENT_FORWARD_TIMEOUT = 5
 
 
 @app.on_event("startup")
@@ -74,6 +79,244 @@ async def redis_errors(
     return responses.JSONResponse(status_code=503)
 
 
+@app.post(
+    "/refresh/{owner_login}/{repo_name}",  # noqa: FS003
+    dependencies=[fastapi.Depends(auth.signature)],
+)
+async def refresh_repo(
+    owner_login: github_types.GitHubLogin,
+    repo_name: github_types.GitHubRepositoryName,
+    redis_cache: utils.RedisCache = fastapi.Depends(  # noqa: B008
+        redis.get_redis_cache
+    ),
+    redis_stream: utils.RedisStream = fastapi.Depends(  # noqa: B008
+        redis.get_redis_stream
+    ),
+) -> responses.Response:
+    installation_json = await github.get_installation_from_login(owner_login)
+    async with github.aget_client(installation_json) as client:
+        try:
+            repository = await client.item(f"/repos/{owner_login}/{repo_name}")
+        except http.HTTPNotFound:
+            return responses.JSONResponse(
+                status_code=404, content="repository not found"
+            )
+
+    await utils.send_repository_refresh(
+        redis_cache, redis_stream, repository, action="user", source="API"
+    )
+    return responses.Response("Refresh queued", status_code=202)
+
+
+RefreshActionSchema = voluptuous.Schema(voluptuous.Any("user", "admin", "internal"))
+
+
+@app.post(
+    "/refresh/{owner_login}/{repo_name}/pull/{pull_request_number}",  # noqa: FS003
+    dependencies=[fastapi.Depends(auth.signature)],
+)
+async def refresh_pull(
+    owner_login: github_types.GitHubLogin,
+    repo_name: github_types.GitHubRepositoryName,
+    pull_request_number: github_types.GitHubPullRequestNumber,
+    action: github_types.GitHubEventRefreshActionType = "user",
+    redis_cache: utils.RedisCache = fastapi.Depends(  # noqa: B008
+        redis.get_redis_cache
+    ),
+    redis_stream: utils.RedisStream = fastapi.Depends(  # noqa: B008
+        redis.get_redis_stream
+    ),
+) -> responses.Response:
+    action = RefreshActionSchema(action)
+
+    installation_json = await github.get_installation_from_login(owner_login)
+    async with github.aget_client(installation_json) as client:
+        try:
+            repository = await client.item(f"/repos/{owner_login}/{repo_name}")
+        except http.HTTPNotFound:
+            return responses.JSONResponse(
+                status_code=404, content="repository not found"
+            )
+
+    await utils.send_pull_refresh(
+        redis_cache,
+        redis_stream,
+        repository,
+        action=action,
+        pull_request_number=pull_request_number,
+        source="API",
+    )
+    return responses.Response("Refresh queued", status_code=202)
+
+
+@app.post(
+    "/refresh/{owner_login}/{repo_name}/branch/{branch}",  # noqa: FS003
+    dependencies=[fastapi.Depends(auth.signature)],
+)
+async def refresh_branch(
+    owner_login: github_types.GitHubLogin,
+    repo_name: github_types.GitHubRepositoryName,
+    branch: str,
+    redis_cache: utils.RedisCache = fastapi.Depends(  # noqa: B008
+        redis.get_redis_cache
+    ),
+    redis_stream: utils.RedisStream = fastapi.Depends(  # noqa: B008
+        redis.get_redis_stream
+    ),
+) -> responses.Response:
+    installation_json = await github.get_installation_from_login(owner_login)
+    async with github.aget_client(installation_json) as client:
+        try:
+            repository = await client.item(f"/repos/{owner_login}/{repo_name}")
+        except http.HTTPNotFound:
+            return responses.JSONResponse(
+                status_code=404, content="repository not found"
+            )
+
+    await utils.send_branch_refresh(
+        redis_cache,
+        redis_stream,
+        repository,
+        action="user",
+        source="API",
+        ref=github_types.GitHubRefType(f"refs/heads/{branch}"),
+    )
+    return responses.Response("Refresh queued", status_code=202)
+
+
+@app.get(
+    "/organization/{owner_id}/usage",  # noqa: FS003
+    dependencies=[fastapi.Depends(auth.signature)],
+)
+async def get_stats(
+    owner_id: github_types.GitHubAccountIdType,
+    redis_cache: utils.RedisCache = fastapi.Depends(  # noqa: B008
+        redis.get_redis_cache
+    ),
+) -> responses.Response:
+    seats = await count_seats.Seats.get(
+        redis_cache, write_users=False, owner_id=owner_id
+    )
+    data = seats.jsonify()
+    if data["organizations"]:
+        if len(data["organizations"]) > 1:
+            raise RuntimeError(
+                "count_seats.Seats.get() returns more than one organization"
+            )
+        repos = data["organizations"][0]["repositories"]
+    else:
+        repos = []
+    return responses.JSONResponse({"repositories": repos})
+
+
+@app.put(
+    "/subscription-cache/{owner_id}",  # noqa: FS003
+    dependencies=[fastapi.Depends(auth.signature)],
+)
+async def subscription_cache_update(
+    owner_id: github_types.GitHubAccountIdType,
+    request: requests.Request,
+    redis_cache: utils.RedisCache = fastapi.Depends(  # noqa: B008
+        redis.get_redis_cache
+    ),
+) -> responses.Response:
+    sub = await request.json()
+    if sub is None:
+        return responses.Response("Empty content", status_code=400)
+    try:
+        await subscription.Subscription.update_subscription(
+            redis_cache, int(owner_id), sub
+        )
+    except NotImplementedError:
+        return responses.Response("Updating subscription is disabled", status_code=400)
+
+    return responses.Response("Cache updated", status_code=200)
+
+
+@app.delete(
+    "/subscription-cache/{owner_id}",  # noqa: FS003
+    dependencies=[fastapi.Depends(auth.signature)],
+)
+async def subscription_cache_delete(
+    owner_id: github_types.GitHubAccountIdType,
+    redis_cache: utils.RedisCache = fastapi.Depends(  # noqa: B008
+        redis.get_redis_cache
+    ),
+) -> responses.Response:
+    try:
+        await subscription.Subscription.delete_subscription(redis_cache, owner_id)
+    except NotImplementedError:
+        return responses.Response("Deleting subscription is disabled", status_code=400)
+    return responses.Response("Cache cleaned", status_code=200)
+
+
+@app.delete(
+    "/tokens-cache/{owner_id}",  # noqa: FS003
+    dependencies=[fastapi.Depends(auth.signature)],
+)
+async def tokens_cache_delete(
+    owner_id: github_types.GitHubAccountIdType,
+    redis_cache: utils.RedisCache = fastapi.Depends(  # noqa: B008
+        redis.get_redis_cache
+    ),
+) -> responses.Response:
+    try:
+        await user_tokens.UserTokens.delete(redis_cache, owner_id)
+    except NotImplementedError:
+        return responses.Response("Deleting tokens is disabled", status_code=400)
+    return responses.Response("Cache cleaned", status_code=200)
+
+
+@app.post("/marketplace", dependencies=[fastapi.Depends(auth.signature)])
+async def marketplace_handler(
+    request: requests.Request,
+    redis_cache: utils.RedisCache = fastapi.Depends(  # noqa: B008
+        redis.get_redis_cache
+    ),
+) -> responses.Response:
+    event_type = request.headers.get("X-GitHub-Event")
+    event_id = request.headers.get("X-GitHub-Delivery")
+    data = await request.json()
+
+    LOG.info(
+        "Marketplace event",
+        event_type=event_type,
+        event_id=event_id,
+        sender=data["sender"]["login"],
+        gh_owner=data["marketplace_purchase"]["account"]["login"],
+    )
+
+    await subscription.Subscription.delete_subscription(
+        redis_cache, data["marketplace_purchase"]["account"]["id"]
+    )
+
+    if config.WEBHOOK_MARKETPLACE_FORWARD_URL:
+        raw = await request.body()
+        try:
+            async with http.AsyncClient(timeout=EVENT_FORWARD_TIMEOUT) as client:
+                await client.post(
+                    config.WEBHOOK_MARKETPLACE_FORWARD_URL,
+                    content=raw.decode(),
+                    headers={
+                        "X-GitHub-Event": event_type,
+                        "X-GitHub-Delivery": event_id,
+                        "X-Hub-Signature": request.headers.get("X-Hub-Signature"),
+                        "User-Agent": request.headers.get("User-Agent"),
+                        "Content-Type": request.headers.get("Content-Type"),
+                    },
+                )
+        except httpx.TimeoutException:
+            LOG.warning(
+                "Fail to forward Marketplace event",
+                event_type=event_type,
+                event_id=event_id,
+                sender=data["sender"]["login"],
+                gh_owner=data["marketplace_purchase"]["account"]["login"],
+            )
+
+    return responses.Response("Event queued", status_code=202)
+
+
 @app.get(
     "/queues/{owner_id}",  # noqa: FS003
     dependencies=[fastapi.Depends(auth.signature_or_token)],
@@ -85,12 +328,12 @@ async def queues(
         redis.get_redis_cache
     ),
 ) -> responses.Response:
-    auth: typing.Optional[github_client.GithubTokenAuth]
+    auth: typing.Optional[github.GithubTokenAuth]
     token = request.headers.get("Authorization")
     if token:
         token = token[6:]  # Drop 'token '
-        auth = github_client.GithubTokenAuth(token)
-        async with github_client.AsyncGithubInstallationClient(auth=auth) as client:
+        auth = github.GithubTokenAuth(token)
+        async with github.AsyncGithubInstallationClient(auth=auth) as client:
             # Check this token as access to this organization
             try:
                 await client.item(f"/user/{owner_id}")
@@ -108,7 +351,7 @@ async def queues(
         queue_type, _, repo_id, branch = queue.split("~")
         if auth is not None:
             # Check this token as access to this repository
-            async with github_client.AsyncGithubInstallationClient(auth=auth) as client:
+            async with github.AsyncGithubInstallationClient(auth=auth) as client:
                 try:
                     await client.item(f"/repositories/{repo_id}")
                 except (http.HTTPNotFound, http.HTTPForbidden, http.HTTPUnauthorized):
@@ -129,6 +372,61 @@ async def queues(
             ] + [int(wp[0]) for wp in train["waiting_pulls"]]
 
     return responses.JSONResponse(status_code=200, content=queues)
+
+
+@app.post("/event", dependencies=[fastapi.Depends(auth.signature)])
+async def event_handler(
+    request: requests.Request,
+    redis_cache: utils.RedisCache = fastapi.Depends(  # noqa: B008
+        redis.get_redis_cache
+    ),
+    redis_stream: utils.RedisStream = fastapi.Depends(  # noqa: B008
+        redis.get_redis_stream
+    ),
+) -> responses.Response:
+    event_type = request.headers.get("X-GitHub-Event")
+    event_id = request.headers.get("X-GitHub-Delivery")
+    data = await request.json()
+
+    try:
+        await github_events.filter_and_dispatch(
+            redis_cache, redis_stream, event_type, event_id, data
+        )
+    except github_events.IgnoredEvent as ie:
+        status_code = 200
+        reason = f"Event ignored: {ie.reason}"
+    else:
+        status_code = 202
+        reason = "Event queued"
+
+    if (
+        config.WEBHOOK_APP_FORWARD_URL
+        and config.WEBHOOK_FORWARD_EVENT_TYPES is not None
+        and event_type in config.WEBHOOK_FORWARD_EVENT_TYPES
+    ):
+        raw = await request.body()
+        try:
+            async with http.AsyncClient(timeout=EVENT_FORWARD_TIMEOUT) as client:
+                await client.post(
+                    config.WEBHOOK_APP_FORWARD_URL,
+                    content=raw.decode(),
+                    headers={
+                        "X-GitHub-Event": event_type,
+                        "X-GitHub-Delivery": event_id,
+                        "X-Hub-Signature": request.headers.get("X-Hub-Signature"),
+                        "User-Agent": request.headers.get("User-Agent"),
+                        "Content-Type": request.headers.get("Content-Type"),
+                    },
+                )
+        except httpx.TimeoutException:
+            LOG.warning(
+                "Fail to forward GitHub event",
+                event_type=event_type,
+                event_id=event_id,
+                sender=data["sender"]["login"],
+            )
+
+    return responses.Response(reason, status_code=status_code)
 
 
 @app.get("/")
